refactor enum code
